Update version name to v0.4.0 (#540)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v           ✰  Thanks for creating a PR! You're awesome! ✰
v Please note that maintainers will only review those PRs with a
completed PR template.
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Purpose of Changes and their Description

Update version name to v0.4.0 when appropriate. 

## Link(s) to Ticket(s) or Issue(s) resolved by this PR


https://linear.app/alloralabs/issue/PROTO-2238/update-the-version-name-string

## Are these changes tested and documented?

Tested using existing tests.

## Still Left Todo

*Fill this out if this is a Draft PR so others can help.*

+++ x/emissions/keeper/actor_utils/util_sort.go
@@ -89,7 +89,7 @@ func FindTopNByScoreDesc(ctx sdk.Context, n uint64, scoresByActor map[Actor]Scor
 
 	topN := make([]Actor, 0)
 	topNBool := make(map[string]bool)
-	for i := 0; i < int(n); i++ {
+	for i := uint64(0); i < n; i++ {
 		if queue.Len() == 0 {
 			break
 		}
